Add IndexController class and configuration classes. Modifying all dtos (Overwrite equals and hashCode methods). Create mappers and services implementations.

+++ src/main/java/com/site/plazam/service/impl/PictureServiceImpl.java
@@ -76,12 +76,14 @@ public class PictureServiceImpl implements PictureService {
 
     @Override
     public void delete(PictureDTO picture, Class className) {
-        if (className == ActorPicture.class) {
-            apr.deleteById(picture.getId());
-        } else if (className == MoviePicture.class) {
-            mpr.deleteById(picture.getId());
-        } else if (className == UserPicture.class) {
-            upr.deleteById(picture.getId());
+        if (picture.getId() != null) {
+            if (className == ActorPicture.class) {
+                apr.deleteById(picture.getId());
+            } else if (className == MoviePicture.class) {
+                mpr.deleteById(picture.getId());
+            } else if (className == UserPicture.class) {
+                upr.deleteById(picture.getId());
+            }
         }
     }
